+++ java example/SimpleLogginApp/.idea/.gitignore
