+++ java example/SimpleLogginApp/pom.xml
